+++ runtime/src/integration/proposals/mod.rs
@@ -1,5 +1,7 @@
 mod council_origin_validator;
 mod membership_origin_validator;
+mod staking_events_handler;
 
 pub use council_origin_validator::CouncilManager;
 pub use membership_origin_validator::{MemberId, MembershipOriginValidator};
+pub use staking_events_handler::StakingEventsHandler;
